requirements.txt
numpy==1.24.3
streamlit==1.27.2
Pillow==9.4.0
keras==2.14.0
tensorflow==2.14.0